scripts_ray/requirements.txt
# Useful Scripts Collection - Python Dependencies
# 用于文档转换、文件处理和系统自动化的实用脚本集合
# 版本: 2.0.0 (重构版)

# ===== 核心数据处理库（必需） =====
pandas>=1.5.0,<3.0.0
openpyxl>=3.0.0,<4.0.0

# ===== Office 文档处理（必需） =====
python-docx>=0.8.11,<2.0.0
python-pptx>=0.6.21,<1.0.0

# ===== 文档转换工具（必需） =====
markitdown>=0.1.0
tiktoken>=0.5.0,<1.0.0

# ===== 文件监控（ext2bind.py 需要） =====
watchdog>=3.0.0,<4.0.0

# ===== 字符编码检测（推荐） =====
# 用于处理各种编码的CSV文件和文本文件
chardet>=5.0.0,<6.0.0

# ===== 系统工具（可选） =====
# 用于系统监控和调试，不是所有脚本都需要
psutil>=5.9.0,<6.0.0

# ===== 网络请求（可选） =====
# 某些脚本可能需要，如果遇到网络相关功能
requests>=2.28.0,<3.0.0

# ===== 开发工具（可选） =====
# 用于代码质量检查，仅开发时使用
# bandit>=1.7.0
# safety>=3.0.0

# ===== 脚本对应的依赖说明 =====
# 
# convert_wmf.py: 无额外依赖 (使用系统 LibreOffice)
# csv2xls.py: pandas, openpyxl, chardet
# ext_img_dp.py: python-docx, python-pptx
# ext_tab_dp.py: python-docx, python-pptx, pandas, openpyxl
# ext2bind.py: watchdog
# gettoken.py: tiktoken
# pptx2md.py: python-pptx
# splitsheets.py: pandas, openpyxl
# 
# 其他脚本: 主要使用标准库

# ===== 安装方法 =====
#
# 安装全部依赖:
# pip install -r requirements.txt
#
# 安装核心依赖（推荐）:
# pip install pandas openpyxl python-docx python-pptx markitdown tiktoken chardet
#
# 使用 conda:
# conda install pandas openpyxl
# pip install python-docx python-pptx markitdown tiktoken chardet watchdog
#
# 分组安装:
# 文档处理: pip install pandas openpyxl python-docx python-pptx markitdown
# 编码检测: pip install chardet
# 文件监控: pip install watchdog
# Token计算: pip install tiktoken 